rename: Rename `Thunder2` to `Thunder`
